rationnalize js strategies
