+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Application/View/Content/ContentListView.js
@@ -2,9 +2,9 @@ import AbstractDataTableView       from '../../../Service/DataTable/View/Abstrac
 import UrlPaginateViewMixin        from '../../../Service/DataTable/Mixin/UrlPaginateViewMixin'
 import DeleteCheckboxListViewMixin from '../../../Service/DataTable/Mixin/DeleteCheckboxListViewMixin'
 import DuplicateIconListViewMixin  from '../../../Service/DataTable/Mixin/DuplicateIconListViewMixin'
-import CellFormatterManager        from '../../../Service/CellFormatter/Manager'
-import BooleanFormatter            from '../../../Service/CellFormatter/BooleanFormatter'
-import DateFormatter               from '../../../Service/CellFormatter/DateFormatter'
+import CellFormatterManager        from '../../../Service/DataFormatter/Manager'
+import BooleanFormatter            from '../../../Service/DataFormatter/BooleanFormatter'
+import DateFormatter               from '../../../Service/DataFormatter/DateFormatter'
 
 /**
  * @class ContentListView
