--- BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Service/CellFormatter/DateFormatter.js
@@ -1,18 +1,0 @@
-import AbstractCellFormatter from './AbstractCellFormatter'
-import DateDataFormatter     from '../DataFormatter/DateFormatter'
-
-/**
- * @class DateFormatter
- */
-class DateFormatter extends AbstractCellFormatter
-{
-    /**
-     * @inheritdoc
-     */
-    getDataFormatter() {
-        return DateDataFormatter;
-    }
-}
-
-// unique instance of DateFormatter
-export default (new DateFormatter);
